+++ doc/source/changelog.rst
@@ -1,13 +1,19 @@
 :Author: Jérôme Kieffer
-:Date: 22/03/2023
+:Date: 16/05/2023
 :Keywords: changelog
 
 Change-log of versions
 ======================
 
-2023.XX dd/mm/2023
+2023.05 dd/mm/2023
 ------------------
 - New detector: Jungfrau 8M used at SwissFEL
+- new method in `Geometry` to guess the number of bin (and avoid oversampling)
+- new method in `AzimuthalIntegrator` to guess the polarization factor
+- CSC-integrators (1D + 2D) implemented in pure python using scipy.sparse
+- Update documentation
+- Minor correction (remove debugging)
+- Packaging for debian12 using meson-python
 
 2023.03 22/03/2023
 ------------------